INSA's member workflow

+++ src/main/java/com/entities/Notification.java
@@ -11,7 +11,7 @@ public class Notification {
 
     @Id
     @GeneratedValue
-    private Integer id;
+    private int id;
 
     @Temporal(TemporalType.TIMESTAMP)
     private Date eventDate;
@@ -27,7 +27,7 @@ public class Notification {
     private Application application;
 
 
-    public long getId() {
+    public int getId() {
         return id;
     }
